requirements.txt
pandas
numpy
matplotlib
spacy
xgboost
scikit-learn
jupyter
python-dotenv